refactor(permissions): extract privacy-permission reads into their own crate
Permission *status* is toolkit-free and is not a property of the settings
window, but it lived in `openlogi-desktop/src/platform/permissions.rs`, so
nothing outside the desktop app could read it without pulling in the whole
widget kit.

The new crate keeps the asymmetry that makes ready-made permission crates
unusable here: it exposes only the non-prompting half of each API pair, plus
`open_pane`. A prompt raised from the wrong process records the grant against
the wrong code-signing identity (#214), so prompting stays with whoever owns
the resource. `request_camera_access` therefore stays in the desktop crate —
it is gpui polling glue, not a permission read.

Split per platform: `macos.rs` (`IOHIDCheckAccess`, `+[CBManager
authorization]`, the System-Settings deep links), `linux.rs` (uinput / hidraw
probes), `tests.rs` (the pure `classify` cases). `lib.rs` is the crate docs,
the two shared types, and the re-export listing.
