+++ crates/openlogi-desktop/Cargo.toml
@@ -23,6 +23,7 @@ openlogi-camera = { path = "../openlogi-camera" }
 openlogi-hook = { path = "../openlogi-hook" }
 openlogi-assets = { path = "../openlogi-assets" }
 openlogi-ipc = { path = "../openlogi-ipc" }
+openlogi-permissions = { path = "../openlogi-permissions" }
 tarpc = { workspace = true }
 gpui = { workspace = true }
 gpui_platform = { workspace = true }
