+++ crates/openlogi-desktop/src/platform/mod.rs
@@ -1,5 +1,4 @@
 //! Platform and OS integration helpers.
 
 pub mod os;
-pub mod permissions;
 pub mod updater;
