+++ release-plz.toml
@@ -100,6 +100,11 @@ release = false
 publish = false
 
 [[package]]
+name = "openlogi-permissions"
+release = false
+publish = false
+
+[[package]]
 name = "openlogi-overlay"
 release = false
 publish = false
